requirements.txt
bcrypt==3.2.0
bidict==0.21.2
cachelib==0.1.1
cffi==1.14.5
click==8.0.1
colorama==0.4.4
dnspython==1.16.0
Flask==2.0.1
Flask-PyMongo==2.3.0
Flask-Session==0.3.2
Flask-SocketIO==5.1.0
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
pycparser==2.20
pymongo==3.11.4
python-dotenv==0.17.1
python-engineio==4.2.0
python-socketio==5.3.0
six==1.16.0
Werkzeug==2.0.1
